yap-macOS:svncom
git-svn-id: https://subversive.cims.nyu.edu/exact@792 a02fd672-c32b-43fa-a32f-4ef1cc386a6b

+++ trunk/progs/robotics/triangle/QuadTree.h
@@ -88,7 +88,7 @@ public:
 	{
 		std::queue<Box*> q;
 		
-		for (int i = 0; i < allLeaf.size(); ++i)
+		for (int i = 0; i < (int)allLeaf.size(); ++i)
 		{
 			if ( allLeaf[i]->contains(x, y, a) )
 			{
